refactor: extract NumberInput, Vec3Input, panelStyles to @gseurat/ui-kit
Moves the drag-to-scrub NumberInput, Vec3Input (with dual layout modes),
and panelStyles from Bricklayer into the shared ui-kit package. Bricklayer
local copies are replaced with re-exports from @gseurat/ui-kit.
